First pass at middleman 3.0 support.
