+++ lib/middleman_extension.rb
